requirements.txt
pytest
requests==2.31.0
selene>=2.0.0rc6
allure-pytest==2.13.5